Improve hnet configuration options handling
setup-conf.sh is installed in /etc/uci-defaults/30_hnet-conf. It reads
its config from /usr/share/hnet/setupconf where configuration variable
are stored (To enable dnsmasq disabling and custom hostname).

When enabled, board specific network script is installed in
/usr/share/hnet/net-conf.sh and then linked to
/etc/uci-defaults/20_hnet-net (so that a user can decide
to install hnet network conf. later). It makes use of uci in
order to change the network configuration.

When installed as a package, both scripts are run (if existing) and then
deleted.